==> BTC/06.Predictions.ipynb <==
# imports
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns

import gc

matrix_array = np.load('Data/matrix_array_20_normalized.npy')
answer_array = np.load('Data/answer_array_20.npy')

matrix_array.shape, answer_array.shape    

from sklearn.model_selection import train_test_split

X = matrix_array
y = answer_array


X_train, X_temp, y_train, y_temp = train_test_split(X,y, test_size=0.8, random_state=1, stratify=y)
X_valid, X_test, y_valid, y_test = train_test_split(X_temp, y_temp, test_size=0.5, random_state=1, stratify=y_temp)

del matrix_array, answer_array, X, y, X_train ,X_temp, y_train, y_temp, X_valid, y_valid
gc.collect()

from datetime import datetime, timezone

# Convert ISO 8601 date strings to Unix timestamp (milliseconds)
def iso_to_unix(iso_str):
    dt = datetime.strptime(iso_str, "%Y-%m-%dT%H:%M:%SZ").replace(tzinfo=timezone.utc)
    return int(dt.timestamp() * 1000)  # Convert to milliseconds

def unix_to_iso(unix_timestamp_ms):
    # Convert milliseconds to seconds
    unix_timestamp_s = unix_timestamp_ms / 1000
    # Create a datetime object from the Unix timestamp
    dt = datetime.utcfromtimestamp(unix_timestamp_s)
    # Format the datetime object as an ISO 8601 date string
    iso_str = dt.strftime("%Y-%m-%dT%H:%M:%SZ")
    return iso_str

# data = pd.read_csv('Data/historical_BTCUSDT_3min_data.csv')
# data_2023 = data[data['Open time']> iso_to_unix('2023-01-01T00:00:00Z')]
# data_2023.to_csv('Data/2023_BTC_data.csv', index=False)
# data_2023 = data_2023.drop(['Close time', 'Ignore'], axis=1)
# data_2023_no_opentime = data_2023.drop(['Open time'], axis=1)
# data_2023

data_2023 = pd.read_csv('Data/until_2023_BTCUSDT_3min_data.csv')
data_2023 = data_2023.drop(['Ignore', 'Close time'], axis=1)
data_2023_no_opentime = data_2023.drop(['Open time'], axis=1)

data_2024 = pd.read_csv('Data/2024_BTCUSDT_3min_data.csv')
data_2024 = data_2024.drop(['Ignore', 'Close time'], axis=1)
data_2024_no_opentime = data_2024.drop(['Open time'], axis=1)

# Define the window sizes for the moving averages
windows = [5, 10, 20, 30, 60, 120, 240]
# Define the window size and standard deviation multiplier for the Bollinger Bands
BB_window_size = 90
BB_std_multiplier = 1


# ADDING MA
for window in windows:
    # Calculate the moving average
    moving_avg = data_2023['Close'].rolling(window).mean()
    # Add the moving average as a new column to the data_2023
    data_2023[f'MA{window}'] = moving_avg

# Calculate the moving average and standard deviation
BB_moving_avg = data_2023['Close'].rolling(BB_window_size).mean()
BB_std_dev = data_2023['Close'].rolling(BB_window_size).std()

# ADDING BB
# Calculate the Bollinger Bands
BB_upper_band = BB_moving_avg + (BB_std_multiplier * BB_std_dev)
BB_lower_band = BB_moving_avg - (BB_std_multiplier * BB_std_dev)

# Add the Bollinger Bands and moving average as new columns to the data_2023
data_2023[f'MA{BB_window_size}'] = BB_moving_avg
data_2023[f'Upper_Band{BB_window_size}'] = BB_upper_band
data_2023[f'Lower_Band{BB_window_size}'] = BB_lower_band





# drop NaN values
data_2023 = data_2023.dropna().reset_index(drop=True)
data_2023

# Define the window sizes for the moving averages
windows = [5, 10, 20, 30, 60, 120, 240]
# Define the window size and standard deviation multiplier for the Bollinger Bands
BB_window_size = 90
BB_std_multiplier = 1


# ADDING MA
for window in windows:
    # Calculate the moving average
    moving_avg = data_2024['Close'].rolling(window).mean()
    # Add the moving average as a new column to the data_2023
    data_2024[f'MA{window}'] = moving_avg

# Calculate the moving average and standard deviation
BB_moving_avg = data_2024['Close'].rolling(BB_window_size).mean()
BB_std_dev = data_2024['Close'].rolling(BB_window_size).std()

# ADDING BB
# Calculate the Bollinger Bands
BB_upper_band = BB_moving_avg + (BB_std_multiplier * BB_std_dev)
BB_lower_band = BB_moving_avg - (BB_std_multiplier * BB_std_dev)

# Add the Bollinger Bands and moving average as new columns to the data_2023
data_2024[f'MA{BB_window_size}'] = BB_moving_avg
data_2024[f'Upper_Band{BB_window_size}'] = BB_upper_band
data_2024[f'Lower_Band{BB_window_size}'] = BB_lower_band





# drop NaN values
data_2024 = data_2024.dropna().reset_index(drop=True)
data_2024

# Define the window sizes for the moving averages
windows = [5, 10, 20, 30, 60, 120, 240]
# Define the window size and standard deviation multiplier for the Bollinger Bands
BB_window_size = 90
BB_std_multiplier = 1


# ADDING MA
for window in windows:
    # Calculate the moving average
    moving_avg = data_2023_no_opentime['Close'].rolling(window).mean()
    # Add the moving average as a new column to the data_2023_no_opentime
    data_2023_no_opentime[f'MA{window}'] = moving_avg

# Calculate the moving average and standard deviation
BB_moving_avg = data_2023_no_opentime['Close'].rolling(BB_window_size).mean()
BB_std_dev = data_2023_no_opentime['Close'].rolling(BB_window_size).std()

# ADDING BB
# Calculate the Bollinger Bands
BB_upper_band = BB_moving_avg + (BB_std_multiplier * BB_std_dev)
BB_lower_band = BB_moving_avg - (BB_std_multiplier * BB_std_dev)

# Add the Bollinger Bands and moving average as new columns to the data_2023_no_opentime
data_2023_no_opentime[f'MA{BB_window_size}'] = BB_moving_avg
data_2023_no_opentime[f'Upper_Band{BB_window_size}'] = BB_upper_band
data_2023_no_opentime[f'Lower_Band{BB_window_size}'] = BB_lower_band





# drop NaN values
data_2023_no_opentime = data_2023_no_opentime.dropna().reset_index(drop=True)
data_2023_no_opentime.columns

# Define the window sizes for the moving averages
windows = [5, 10, 20, 30, 60, 120, 240]
# Define the window size and standard deviation multiplier for the Bollinger Bands
BB_window_size = 90
BB_std_multiplier = 1


# ADDING MA
for window in windows:
    # Calculate the moving average
    moving_avg = data_2024_no_opentime['Close'].rolling(window).mean()
    # Add the moving average as a new column to the data_2023_no_opentime
    data_2024_no_opentime[f'MA{window}'] = moving_avg

# Calculate the moving average and standard deviation
BB_moving_avg = data_2024_no_opentime['Close'].rolling(BB_window_size).mean()
BB_std_dev = data_2024_no_opentime['Close'].rolling(BB_window_size).std()

# ADDING BB
# Calculate the Bollinger Bands
BB_upper_band = BB_moving_avg + (BB_std_multiplier * BB_std_dev)
BB_lower_band = BB_moving_avg - (BB_std_multiplier * BB_std_dev)

# Add the Bollinger Bands and moving average as new columns to the data_2023_no_opentime
data_2024_no_opentime[f'MA{BB_window_size}'] = BB_moving_avg
data_2024_no_opentime[f'Upper_Band{BB_window_size}'] = BB_upper_band
data_2024_no_opentime[f'Lower_Band{BB_window_size}'] = BB_lower_band





# drop NaN values
data_2024_no_opentime = data_2024_no_opentime.dropna().reset_index(drop=True)
data_2024_no_opentime.columns

# Initialize empty lists to hold the data
matrix_list = []


for i in range(len(data_2023) - 19):  # Ensure there are 20 rows available
    # Create a matrix of 20 rows
    matrix = data_2023.iloc[i:i+20].values
    matrix_list.append(matrix)

# Convert the lists to numpy arrays
data_2023_matrix_20 = np.array(matrix_list)


matrix_list = []


for i in range(len(data_2024) - 19):  # Ensure there are 20 rows available
    # Create a matrix of 20 rows
    matrix = data_2024.iloc[i:i+20].values
    matrix_list.append(matrix)

# Convert the lists to numpy arrays
data_2024_matrix_20 = np.array(matrix_list)

# Initialize empty lists to hold the data
matrix_list = []


for i in range(len(data_2023_no_opentime) - 19):  # Ensure there are 20 rows available
    # Create a matrix of 20 rows
    matrix = data_2023_no_opentime.iloc[i:i+20].values
    matrix_list.append(matrix)

# Convert the lists to numpy arrays
data_2023_no_opentime_matrix_20 = np.array(matrix_list)


# Initialize empty lists to hold the data
matrix_list = []


for i in range(len(data_2024_no_opentime) - 19):  # Ensure there are 20 rows available
    # Create a matrix of 20 row
    matrix = data_2024_no_opentime.iloc[i:i+20].values
    matrix_list.append(matrix)

# Convert the lists to numpy arrays
data_2024_no_opentime_matrix_20 = np.array(matrix_list)


data_2023_matrix_20.shape, data_2023_no_opentime_matrix_20.shape, data_2024_matrix_20.shape, data_2024_no_opentime_matrix_20.shape

np.save('Data/data_2023_matrix_20.npy', data_2023_matrix_20)
np.save('Data/data_2024_matrix_20.npy', data_2024_matrix_20)

## RFC model

import joblib

# Assuming your model is named RFC
RFC = joblib.load('Models/RFC_model_rank_1.pkl')
scaler = joblib.load('Scalers/StandardScaler_20.pkl')

# scale the data

# Assuming data is your 600k matrices concatenated into a single 3D numpy array of shape (600000, 20, 19)
temp = data_2023_no_opentime_matrix_20.reshape(-1, 19)  # Reshape to 2D for standardization
temp_normalized = scaler.transform(temp)

# Reshape back to 3D
data_2023_matrix_20_normalized = temp_normalized.reshape(-1, 20, 19)

data_2023_matrix_20_normalized_flattened = np.array([x.flatten() for x in data_2023_matrix_20_normalized])

y_pred = RFC.predict(data_2023_matrix_20_normalized_flattened)

np.save('Data/RFC_20_2023_prediction.npy', y_pred)

# scale the data

# Assuming data is your 600k matrices concatenated into a single 3D numpy array of shape (600000, 20, 19)
temp = data_2024_no_opentime_matrix_20.reshape(-1, 19)  # Reshape to 2D for standardization
temp_normalized = scaler.transform(temp)

# Reshape back to 3D
data_2024_matrix_20_normalized = temp_normalized.reshape(-1, 20, 19)

data_2024_matrix_20_normalized_flattened = np.array([x.flatten() for x in data_2024_matrix_20_normalized])

y_pred = RFC.predict(data_2024_matrix_20_normalized_flattened)

np.save('Data/RFC_20_2024_prediction.npy', y_pred)

## LSTM model

# imports
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns

import torch
import torch.nn as nn

from torch.utils.data import TensorDataset, DataLoader
import torch.optim as optim
from torch.optim.lr_scheduler import ReduceLROnPlateau
from torchsummary import summary
from tqdm import  tqdm

import os, json, gc, io, joblib
from contextlib import redirect_stdout

from sklearn.model_selection import train_test_split

# Custom LSTM model
class LSTMModel(nn.Module):
    def __init__(self, input_dim, hidden_dim, output_dim):
        super(LSTMModel, self).__init__()

        # Multi-layer LSTM
        self.lstm = nn.LSTM(input_dim, hidden_dim, num_layers=3, batch_first=True, dropout=0.2)

        # Flatten layer 

        # Define Linear layers with Batch Normalization, GELU, and Dropout
        self.linear1 = nn.Linear(hidden_dim, 64)
        self.batch_norm1 = nn.BatchNorm1d(64)
        self.gelu1 = nn.GELU()
        self.dropout1 = nn.Dropout(0.2)

        self.linear2 = nn.Linear(64, 32)
        self.batch_norm2 = nn.BatchNorm1d(32)
        self.gelu2 = nn.GELU()
        self.dropout2 = nn.Dropout(0.2)

        self.linear3 = nn.Linear(32, 16)
        self.batch_norm3 = nn.BatchNorm1d(16)
        self.gelu3 = nn.GELU()

        # Output layer
        self.output_layer = nn.Linear(16, output_dim)

    def forward(self, x):
        # Pass input through LSTM layers
        # print("Output type1:", type(x))
        (lstm_out, _) = self.lstm(x)

        # Taking the output of the last time step
        # print("Output type2:", type(x))
        x = lstm_out[:, -1, :]

        # Pass through Linear layers
        x = self.linear1(x)
        x = self.batch_norm1(x)
        x = self.gelu1(x)
        x = self.dropout1(x)

        x = self.linear2(x)
        x = self.batch_norm2(x)
        x = self.gelu2(x)
        x = self.dropout2(x)

        x = self.linear3(x)
        x = self.batch_norm3(x)
        x = self.gelu3(x)

        # Output layer
        x = self.output_layer(x)
        
        return x 


# Initialize the model
model = LSTMModel(input_dim=19, hidden_dim=128, output_dim=3)

state_dict = torch.load('Models/LSTM/LSTM_model_state_dict_custominitialize.pth')
model.load_state_dict(state_dict)

# Setting device to GPU if available
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
# device = torch.device('cpu')
model = model.to(device)

import joblib

scaler = joblib.load('Scalers/StandardScaler_20.pkl')

# scale the data

# Assuming data is your 600k matrices concatenated into a single 3D numpy array of shape (600000, 20, 19)
temp = data_2024_no_opentime_matrix_20.reshape(-1, 19)  # Reshape to 2D for standardization
temp_normalized = scaler.transform(temp)

# Reshape back to 3D
data_2024_matrix_20_normalized = temp_normalized.reshape(-1, 20, 19)

data_2024_matrix_20_normalized.shape

X = data_2024_matrix_20_normalized

X_tensor = torch.tensor(X, dtype=torch.float32)


import torch
import pandas as pd

# Ensure the model is in evaluation mode
model.eval()

# Move tensor to the same device as model
X_tensor = X_tensor.to(device)

# No need to track gradients for prediction
with torch.no_grad():
    # Predict
    predictions = model(X_tensor)

    # Apply softmax to convert logits to probabilities
    probabilities = torch.softmax(predictions, dim=1)  # Assuming output_dim corresponds to class counts
    # If you need class labels instead of probabilities:
    predicted_labels = torch.argmax(predictions, dim=1)

# Move predictions back to cpu for converting to numpy array
# probabilities = probabilities.cpu().numpy()
predicted_labels = predicted_labels.cpu().numpy()  # Uncomment if using argmax
print(predicted_labels)

np.save('Data/LSTM_2024_predictions.npy', predicted_labels)
